requirements.txt
flask==2.3.2
flask-restful==0.3.10
pykwalify==1.8.0
GitPython==3.1.31
pyyaml==6.0
rich==13.4.2
pandas==2.0.2
litellm==1.75.2
jinja2==3.1.2

